+++ src/main/java/me/pajic/accessorify/accessories/ElytraAccessory.java
@@ -1,7 +1,11 @@
 package me.pajic.accessorify.accessories;
 
+import io.wispforest.accessories.api.attributes.AccessoryAttributeBuilder;
 import io.wispforest.accessories.api.slot.SlotReference;
 import me.pajic.accessorify.util.MultiVersionUtil;
+import me.pajic.accessorify.util.compat.AileronCompat;
+import me.pajic.accessorify.util.compat.CompatFlags;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 //? if <= 1.21.1
@@ -23,6 +27,12 @@ public class ElytraAccessory implements SlotCopyingAccessory {
     @Override
     public String getSlot() {
         return "cape";
+    }
+
+    @SuppressWarnings("resource")
+    @Override
+    public void getDynamicModifiers(ItemStack stack, SlotReference reference, AccessoryAttributeBuilder builder) {
+        if (CompatFlags.AILERON_LOADED) AileronCompat.addModifiers(builder, stack, reference.entity().level().registryAccess().lookupOrThrow(Registries.ENCHANTMENT));
     }
 
     @Override